ecommerce_app/requirements.txt
Django>=6.0.2
mysqlclient>=2.2.7
djangorestframework>=3.16.1
djangorestframework-xml>=2.0.0
tweepy>=4.16.0
requests-oauthlib>=2.0.0
